Create index.html
We developed a web application to solve the problems of insecurity residentials starting from estates in nigeria.
There is a registration form that collects the data of all occupants of the estates and store it in a database for verification before pass is granted. If your name for instance cannot be found in the database then you have to answer some security questions before a pass is granted.